requirements.txt
sentence-transformers
langchain
flask
pypdf
notebook
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain-google-genai
google-generativeai
langchain_experimental
pyreadline

-e .